+++ sample/proto_v2/arr_in_arr_cfg.lua
@@ -3,10 +3,10 @@
 return {
   [1] = {
     count = 3,
-    data_ver = "2.8.0.20200609105146",
+    data_ver = "1.0.0.0",
     description = "Test arr_in_arr_cfg",
     hash_code = "no hash code",
-    xres_ver = "2.8.0",
+    xres_ver = "2.8.1",
   },
   [2] = "arr_in_arr_cfg",
   arr_in_arr_cfg = {