+++ sample/proto_v2/kind_option.mod.lua
@@ -266,28 +266,6 @@ local const_res = {
       package = "",
       path = "kind.proto",
     },
-    {
-      name = "xresloader.proto",
-      options = {
-        optimize_for = {
-          name = "SPEED",
-          number = 1,
-        },
-      },
-      package = "org.xresloader",
-      path = "xresloader.proto",
-    },
-    {
-      name = "xresloader_ue.proto",
-      options = {
-        optimize_for = {
-          name = "SPEED",
-          number = 1,
-        },
-      },
-      package = "org.xresloader.ue",
-      path = "xresloader_ue.proto",
-    },
   },
 }